Keep the terminal UI runtime out of the CLI startup path
`output.ts` and `error.ts` imported `tokenItemToString` from
`private/node/ui/components/TokenizedText.tsx`, and `error.ts` imported
`ClientError` from graphql-request for a single `instanceof` check.
TokenizedText is a React component module, so those two imports pulled
react-reconciler (341 KB), yoga-layout and its WebAssembly (125 KB), ink
and react into the module graph of every command, plus graphql (254 KB),
tr46 (231 KB) and whatwg-url. Practically everything imports output.ts,
so `shopify --version` was loading a terminal UI renderer and a GraphQL
client to print a version string.

Moves the token types, tokenItemToString and appendToTokenItem into
token-item.ts, which imports nothing, and repoints every import site.
TokenizedText.tsx keeps the component and imports the types. Replaces the
`instanceof ClientError` check with a structural match on the same shape;
ClientError is the only error reaching that path carrying both `response`
and `request`, and the cli-kit wrapper carries `statusCode` instead.

No behaviour change. Measured on the bundled CLI, clean rebuilds on both
sides, cachegrind instruction counts under `node --predictable`:

                              before          after      delta
  shopify --version   1,543,699,375    737,632,410     -52.2%
  shopify --help      1,606,749,593    809,187,564     -49.6%
  --version wall              174 ms          99 ms       -43%
  JS loaded at boot          5.31 MB        4.16 MB       -22%

Output is byte-identical for --version, --help, app --help and commands.
