+++ packages/cli-kit/src/private/node/ui/components/DangerousConfirmationPrompt.tsx
@@ -1,5 +1,5 @@
 import {TextInput} from './TextInput.js'
-import {InlineToken, TokenItem, TokenizedText} from './TokenizedText.js'
+import {TokenizedText} from './TokenizedText.js'
 import {InfoTable, InfoTableProps} from './Prompts/InfoTable.js'
 import {handleCtrlC, useComplete} from '../../ui.js'
 import useLayout from '../hooks/use-layout.js'
@@ -11,6 +11,7 @@ import usePrompt, {PromptState} from '../hooks/use-prompt.js'
 import React, {FunctionComponent, useCallback, useEffect, useState} from 'react'
 import {Box, useInput, Text} from 'ink'
 import figures from 'figures'
+import type {InlineToken, TokenItem} from './token-item.js'
 
 export interface DangerousConfirmationPromptProps {
   message: string
