+++ packages/cli-kit/src/private/node/ui/components/List.tsx
@@ -1,6 +1,7 @@
-import {InlineToken, TokenItem, TokenizedText} from './TokenizedText.js'
+import {TokenizedText} from './TokenizedText.js'
 import {Box, Text, TextProps} from 'ink'
 import React, {FunctionComponent} from 'react'
+import type {InlineToken, TokenItem} from './token-item.js'
 
 export interface CustomListItem {
   type?: string
