+++ packages/cli-kit/src/public/node/output.ts
@@ -19,7 +19,7 @@ import {
   RawContentToken,
   SubHeadingContentToken,
 } from '../../private/node/content-tokens.js'
-import {tokenItemToString} from '../../private/node/ui/components/TokenizedText.js'
+import {tokenItemToString} from '../../private/node/ui/components/token-item.js'
 import {consoleLog, consoleWarn, output} from '../../private/node/output.js'
 import stripAnsi from 'strip-ansi'
 import {Writable} from 'stream'
